requirements.txt
FastAPI~=0.115.12
uvicorn[standard]~=0.34.0
requests~=2.32.3
requests-oauthlib==2.0.0
setuptools
pydantic~=2.11.3
typing_extensions~=4.12.2
python-dateutil~=2.9.0.post0
urllib3 >= 1.25.3, < 2.1.0
python-dotenv~=1.1.0
basyx-python-sdk==1.1.0
aas-python-http-client==1.0.2
pycouchdb==1.16.0
paho-mqtt==2.1.0
starlette~=0.46.1
oauthlib~=3.2.2
PyYAML~=6.0.2
pydantic_core~=2.33.1